(INCOMPLETE) moved chunk into world_gen and continued moving stuff into their own files
