+++ main.py
@@ -1,17 +1,11 @@
-from random import randint, seed, choices, randrange
-from perlin_noise import PerlinNoise
 from sys import exit as sysexit
-from math import ceil
-import opensimplex
+from random import seed
 import pygame
-import json
 import os
-import time
 from pygame.locals import  (
-    HWSURFACE, SRCALPHA, DOUBLEBUF,
-    K_w, K_e, K_a, K_d,
-    K_1, K_9, K_0, K_F5,
     MOUSEBUTTONDOWN, KEYDOWN,
+    HWSURFACE, DOUBLEBUF,
+    K_e, K_F5,
     QUIT,
 )
 
@@ -21,7 +15,7 @@ from src.images import *
 from src.player import Player
 from src.particle import Particle
 from src.block import *
-from src.chunk import Chunk
+from src.world_gen import Chunk
 
 def loading():
     font = pygame.font.Font(REGULAR_FONT_LOC, 120)
@@ -36,203 +30,11 @@ pygame.display.set_caption("2D Minecraft")
 loading()
 clock = pygame.time.Clock()
 mixer = pygame.mixer.init()
-noise = opensimplex.OpenSimplex(seed=SEED)
-pnoise = PerlinNoise(seed=SEED)
 seed(SEED)
-
-structures = {}
-for folder in os.listdir(pathof("data/structures/")):
-    structures[folder] = {}
-    for struct in os.listdir(os.path.join(pathof("data/structures/"), folder)):
-        if struct != "distribution.structure":
-            data = open(pathof(f"data/structures/{folder}/{struct}"), "r").readlines()
-            split = data.index("structure:\n")
-            split2 = data.index("origin:\n")
-            legends = {legend.split(":")[0]: legend.split(":")[1] for legend in [legend[:-1] for legend in data[:split]]}
-            structure = [line[:-1] for line in data[split+1:split2]]
-            origin = (int(data[-1].split(" ")[0]), int(data[-1].split(" ")[1]))
-            blocks = {}
-            for y in range(len(structure)):
-                for x in range(len(structure[y])):
-                    try: legends[structure[y][x]]
-                    except: pass
-                    else: blocks[(x-origin[0], y-origin[1])] = legends[structure[y][x]]
-            structures[folder][struct[:-10]] = (origin, blocks)
-        else:
-            distribution = open(pathof(f"data/structures/{folder}/distribution.structure"), "r").readlines()
-            weights = [int(d[:-2].split(" ")[1]) for d in distribution]
-            files = [d[:-2].split(" ")[0] for d in distribution]
-            structures[folder]["distribution"] = {"weights": weights, "files": files}
-
-class StructureGenerator(object):
-    def __init__(self, folder, obstruction=False):
-        self.folder = folder
-        self.obstruction = obstruction
-        self.files = structures[folder]
-        self.distribution = structures[folder]["distribution"]
-        self.block_data = {}
-        
-        max_sizes = []
-        for file in self.files:
-            if file != "distribution":
-                self.block_data[file] = structures[folder][file]
-                max_sizes.append((max(x for x, y in self.block_data[file][1]) - min(x for x, y in self.block_data[file][1]) + 1,
-                            max(y for x, y in self.block_data[file][1]) - min(y for x, y in self.block_data[file][1]) + 1))
-        self.max_size = max(max_sizes)
-        
-        self.chunks_to_check = int(ceil(self.max_size[0] / CHUNK_SIZE)), int(ceil(self.max_size[1] / CHUNK_SIZE))
-
-    def generate(self, origin):
-        seed(SEED + origin[0] * CHUNK_SIZE + origin[1] * CHUNK_SIZE)
-        file = choices(self.distribution["files"], weights=self.distribution["weights"])[0]
-        mirror = randint(0, 1)
-        
-        block_data = {
-            (origin[0] + offset[0] if mirror else origin[0] - offset[0], origin[1] + offset[1]):
-            (block if "," not in block else (choices([i.split("=")[0] for i in block.split(",")],
-            weights=[int(i.split("=")[1]) for i in block.split(",")])[0]))
-            for offset, block in self.block_data[file][1].items()
-        }
-        
-        return block_data
-
-def get_structures(x: int, y: int, generator: StructureGenerator, chance: tuple) -> list:
-    """Get structures inside the current chunk (x, y)
-
-    Args:
-        x (int): chunk position x
-        y (int): chunk position y
-        generator (StructureGenerator): structure object to generate
-        chance (tuple): the odds of the structure generating, (number-of-structure-possible-per-chunk, odds-off-the-structure-generating-per-block-in-chunk)
-
-    Returns:
-        list: a list containing the block data of the structures in the chunk
-    """
-    out = []
-    seed(SEED + x * CHUNK_SIZE + y * CHUNK_SIZE)
-    for _ in range(chance[0]):
-        if randint(0, chance[1]) == 0:
-            block_x = x * CHUNK_SIZE + randrange(0, CHUNK_SIZE)
-            # Generate on the surface of the world
-            grass_y = terrain_generate(block_x)-1
-            
-            # If it is cut off by a cave, don't generate
-            if (92.7 < cave_generate([block_x/70, grass_y/70]) < 100) or (92.7 < cave_generate([block_x/70, (grass_y+1)/70]) < 100):
-                return out
-            # Structures that are not in this chunk
-            if not 0 <= grass_y - y * CHUNK_SIZE < CHUNK_SIZE:
-                return out
-            
-            out.append(generator.generate((block_x, grass_y)))
-    return out
-
-def generate_structures(x: int, y: int, chunk_data: dict, generator: StructureGenerator, chance: tuple) -> dict:
-    """Check the surrounding chunks for structures that generates in the current chunk (x, y), and then returns the chunk data with the structure
-
-    Args:
-        x (int): chunk position x
-        y (int): chunk position y
-        chunk_data (dict): original chunk data
-        generator (StructureGenerator): the structure object to generate
-        chance (tuple): the odds of the structure generating, (number-of-structure-possible-per-chunk, odds-of-a-chunk-containing-the-structure)
-
-    Returns:
-        dict: the chunk data with the structure
-    """
-    chunk_data_orig = chunk_data.copy()
-    # Check all chunks around the current chunk within the size limit of the structure
-    for ox in range(-generator.chunks_to_check[0], generator.chunks_to_check[0] + 1):
-        for oy in range(-generator.chunks_to_check[1], generator.chunks_to_check[1] + 1):
-            # Get the surrounding structures that might protrude into the current chunk
-            structs = get_structures(x + ox, y + oy, generator, chance)
-            for struct in structs:
-                for block, block_name in struct.items():
-                    # If there are parts of that structure inside the current chunk, generate that part in this chunk
-                    if 0 <= block[0]-x*CHUNK_SIZE < CHUNK_SIZE and 0 <= block[1]-y*CHUNK_SIZE < CHUNK_SIZE:
-                        if block in chunk_data:
-                            # Check if that position already has block and whether it can be overwritten by the structure
-                            # i.e. leaves can replace grass but not grass blocks
-                            if "overwriteable" in block_data[block_name]:
-                                if chunk_data[block] in block_data[block_name]["overwriteable"]:
-                                    chunk_data[block] = block_name
-                                else:
-                                    # Obstruction determines whether the struture can generate when there are non-overwriteable blocks in its way
-                                    if generator.obstruction:
-                                        return chunk_data_orig
-                            else:
-                                if not generator.obstruction:
-                                    chunk_data[block] = block_name
-                                else:
-                                    return chunk_data_orig
-                        else:
-                            chunk_data[block] = block_name
-    return chunk_data
-
-def terrain_generate(x: int) -> float:
-    """Takes the x position of a block and returns the height it has to be at"""
-    return -int(noise.noise2(x*0.1, 0)*5)+5
-
-def cave_generate(coords: list) -> float:
-    """Takes the coordinates of a block and returns the noise map value for cave generation"""
-    noise_height = pnoise(coords)
-    noise_height = noise_height + 0.5 if noise_height > 0 else 0
-    noise_height = int(pow(noise_height * 255, 0.9))
-    return noise_height
-
-def generate_chunk(x: int, y: int) -> dict:
-    """Takes the chunk coordinates and returns a dictionary containing the block data inside the chunk"""
-    seed(x*CHUNK_SIZE+y*CHUNK_SIZE)
-    chunk_data = {}
-    for y_pos in range(CHUNK_SIZE):
-        for x_pos in range(CHUNK_SIZE):
-            target = (x * CHUNK_SIZE + x_pos, y * CHUNK_SIZE + y_pos)
-            block_name = ""
-            
-            # Cave noise map
-            cave_noise_map_coords = [target[0]/70, target[1]/70]
-            # Don't generate blocks if it satifies a certain range of values in the cave noise map, AKA a cave
-            cave_noise_map_value = cave_generate(cave_noise_map_coords)
-            
-            if not (92.7 < cave_noise_map_value < 100):
-                # Generate terrain
-                height = terrain_generate(target[0])
-                if target[1] == height:
-                    block_name = "grass_block"
-                elif height+1 <= target[1] < height+4:
-                    block_name = "dirt"
-                elif target[1] >= height+4:
-                    block_name = "stone"
-                if target[1] == height-1:
-                    if not (92.7 < cave_generate([target[0]/70, (target[1]+1)/70]) < 100):
-                        if randint(0, 2) == 0:
-                            block_name = "grass"
-                        if randint(0, 21) == 0:
-                            block_name = choices(["poppy", "dandelion"], weights=[1, 2])[0]
-                if block_name != "":
-                    chunk_data[target] = block_name
-
-    # Generate structures
-    generator = StructureGenerator("oak_tree", obstruction=False)
-    chunk_data = generate_structures(x, y, chunk_data, generator, (1, 2))
-    generator = StructureGenerator("tall_grass", obstruction=True)
-    chunk_data = generate_structures(x, y, chunk_data, generator, (4, 3))
-    return chunk_data
 
 blocks = {}
 chunks = {}
 player = Player()
-player.inventory.add_item("grass_block")
-player.inventory.add_item("dirt")
-player.inventory.add_item("stone")
-player.inventory.add_item("grass")
-player.inventory.add_item("oak_log")
-player.inventory.add_item("oak_leaves")
-player.inventory.add_item("oak_planks")
-player.inventory.add_item("cobblestone")
-player.inventory.add_item("glass")
-player.inventory.add_item("poppy")
-player.inventory.add_item("dandelion")
-player.inventory.add_item("tall_grass")
 
 pygame.mouse.set_visible(False)
 remove_blocks = []
@@ -242,18 +44,13 @@ def draw(screen):
     mpos = VEC(pygame.mouse.get_pos())
     block_pos = (player.pos+(mpos-player.rect.topleft))//BLOCK_SIZE
     for chunk in rendered_chunks:
-        chunks[chunk].render()
+        chunks[chunk].draw()
     for particle in Particle.instances:
         particle.draw(player.camera, screen)
     player.draw(screen)
     
     # Debug stuff
-    if debug:        
-        for chunk in rendered_chunks:
-            chunks[chunk].debug()
-        pygame.draw.rect(screen, (255, 255, 255), player.rect, width=1)
-        for rect in player.detecting_rects:
-            pygame.draw.rect(screen, (255, 0, 0), rect, width=1)
+    if debug:
         screen.blit(text(f"Seed: {SEED}"), (6, 0))
         screen.blit(text(f"Velocity: {(round(player.vel.x, 3), round(player.vel.y, 3))}"), (6, 24))
         screen.blit(text(f"Position: {inttup(player.coords)}"), (6, 48))
@@ -261,6 +58,7 @@ def draw(screen):
         screen.blit(text(f"Chunk: {inttup(player.coords//CHUNK_SIZE)}"), (6, 96))
         screen.blit(text(f"Chunks loaded: {len(chunks)}"), (6, 120))
         screen.blit(text(f"Rendered blocks: {len(blocks)}"), (6, 144))
+        screen.blit(text(f"FPS: {len(blocks)}"), (6, 144))
         if not player.inventory.visible:
             if inttup(block_pos) in blocks:
                 screen.blit(text(f"{blocks[inttup(block_pos)].name.replace('_', ' ')}", color=(255, 255, 255)), (mpos[0]+12, mpos[1]-36))
@@ -270,6 +68,30 @@ def draw(screen):
         player.crosshair.draw(screen, mpos)
 
     pygame.display.flip()
+    
+def debug() -> None:
+    # Draw chunk outline
+    for chunk in rendered_chunks:
+        chunks[chunk].debug()
+    # Draw player rect
+    pygame.draw.rect(screen, (255, 255, 255), player.rect, width=1)
+    # Display the blocks that the player is currently calculating collision with
+    for rect in player.detecting_rects:
+        pygame.draw.rect(screen, (255, 0, 0), rect, width=1)
+
+    debug_values = {
+        "FPS": int(clock.get_fps()),
+        "Seed": SEED, 
+        "Velocity": (round(player.vel.x, 3), round(player.vel.y, 3)),
+        "Positon": inttup(player.coords),
+        "Camera offset": inttup(player.pos-player.camera.pos-VEC(SCR_DIM)/2+player.size/2),
+        "Chunk": inttup(player.coords//CHUNK_SIZE),
+        "Chunks loaded": len(chunks),
+        "Rendered blocks": len(blocks)
+    }
+    
+    for line, name in enumerate(debug_values):
+        screen.blit(text(f"{name}: {debug_values[name]}"), (6, SPACING * line))
 
 running = True
 debug = False
@@ -277,10 +99,9 @@ debug = False
 while running:
     dt = clock.tick_busy_loop(FPS) / 16
     if dt > 12: dt = 12
-    pygame.display.set_caption(f"2D Minecraft | FPS: {int(clock.get_fps())}")
+    pygame.display.set_caption(f"2D Minecraft")
     
     mpos = VEC(pygame.mouse.get_pos())
-
     mouse_state = 0
     for event in pygame.event.get():
         if event.type == QUIT:
@@ -303,8 +124,8 @@ while running:
                         
                 if event.button == 3:
                     if player.holding:
-                        if "counterparts" in block_data[player.holding]:
-                            counterparts = block_data[player.holding]["counterparts"]
+                        if "counterparts" in BLOCK_DATA[player.holding]:
+                            counterparts = BLOCK_DATA[player.holding]["counterparts"]
                             for counterpart in counterparts:
                                 c_pos = VEC(block_pos)+VEC(inttup(counterpart.split(" ")))
                                 c_neighbors = {
@@ -313,11 +134,11 @@ while running:
                                     "-1 0": inttup((c_pos.x-1, c_pos.y)),
                                     "1 0": inttup((c_pos.x+1, c_pos.y))
                                 }
-                                if not is_placeable(c_pos, block_data[counterparts[counterpart]], c_neighbors, c=block_pos):
+                                if not is_placeable(c_pos, BLOCK_DATA[counterparts[counterpart]], c_neighbors, c=block_pos):
                                     break
                             else:
                                 for counterpart in counterparts:
-                                    if not is_placeable(block_pos, block_data[player.holding], neighbors, c=c_pos):
+                                    if not is_placeable(block_pos, BLOCK_DATA[player.holding], neighbors, c=c_pos):
                                         break
                                 else:
                                     set_block(block_pos, player.holding, neighbors)
@@ -331,7 +152,7 @@ while running:
                                         }
                                         set_block(VEC(block_pos)+VEC(inttup(counterpart.split(" "))), counterparts[counterpart], c_neighbors)
                         else:
-                            if is_placeable(block_pos, block_data[player.holding], neighbors):
+                            if is_placeable(block_pos, BLOCK_DATA[player.holding], neighbors):
                                 set_block(block_pos, player.holding, neighbors)
                                 
         if event.type == KEYDOWN:
@@ -391,6 +212,7 @@ while running:
     for particle in Particle.instances:
         particle.update(blocks, dt)
     
+    if debug: debug()
     draw(screen)
 
 pygame.quit()
