+++ src/block.py
@@ -2,14 +2,14 @@ import pygame
 import json
 import os
 
-from constants import *
-from utils import *
-from images import *
-from particle import Particle
+from src.constants import *
+from src.utils import *
+from src.images import *
+from src.particle import Particle
 
-block_data = {}
+BLOCK_DATA = {}
 for j in os.listdir(pathof("data/blocks/")):
-    block_data[j[:-5]] = json.loads(open(os.path.join(pathof("data/blocks/"), j), "r").read())
+    BLOCK_DATA[j[:-5]] = json.loads(open(os.path.join(pathof("data/blocks/"), j), "r").read())
 
 class Block(pygame.sprite.Sprite):
     instances = []
@@ -18,7 +18,7 @@ class Block(pygame.sprite.Sprite):
         pygame.sprite.Sprite.__init__(self)
         self.__class__.instances[tuple(pos)] = self
         self.name = name
-        self.data = block_data[self.name]
+        self.data = BLOCK_DATA[self.name]
         self.chunk = chunk
         self.coords = VEC(pos)
         self.pos = self.coords * BLOCK_SIZE
@@ -37,7 +37,7 @@ class Block(pygame.sprite.Sprite):
 
     def update(self):
         if not is_supported(self.pos, self.data, self.neighbors):
-            remove_blocks.append(self.coords)
+            remove_blocks.append(self.coords) # !
 
     def draw(self, camera, screen):
         self.rect.topleft = self.pos - camera.pos
@@ -46,35 +46,35 @@ class Block(pygame.sprite.Sprite):
 def remove_block(pos, data, neighbors):
     pos = inttup(pos)
     for _ in range(randint(18, 26)):
-        Particle("block", VEC(pos)*BLOCK_SIZE+VEC(randint(0, BLOCK_SIZE), randint(0, BLOCK_SIZE)), master=blocks[pos])
+        Particle("block", VEC(pos)*BLOCK_SIZE+VEC(randint(0, BLOCK_SIZE), randint(0, BLOCK_SIZE)), master=blocks[pos]) # !
     chunk = (pos[0] // CHUNK_SIZE, pos[1] // CHUNK_SIZE)
     if "next_layer" in data:
-        blocks[pos] = Block(chunk, pos, data["next_layer"])
-        chunks[chunk].block_data[pos] = data["next_layer"]
+        blocks[pos] = Block(chunk, pos, data["next_layer"]) # !
+        chunks[chunk].block_data[pos] = data["next_layer"] # !
     else:
-        del blocks[pos]
-        del chunks[chunk].block_data[pos]
+        del blocks[pos] # !
+        del chunks[chunk].block_data[pos] # !
     for neighbor in neighbors:
-        if neighbors[neighbor] in blocks:
-            blocks[neighbors[neighbor]].update()
+        if neighbors[neighbor] in blocks: # !
+            blocks[neighbors[neighbor]].update() # !
 
 def set_block(pos, name, neighbors):
     pos = inttup(pos)
     chunk = (pos[0] // CHUNK_SIZE, pos[1] // CHUNK_SIZE)
     try:
-        blocks[pos] = Block(chunks[chunk], pos, name)
-        chunks[chunk].block_data[pos] = name
+        blocks[pos] = Block(chunks[chunk], pos, name) # !
+        chunks[chunk].block_data[pos] = name # !
     except: pass
     for neighbor in neighbors:
         chunk = (neighbors[neighbor][0] // CHUNK_SIZE, neighbors[neighbor][1] // CHUNK_SIZE)
-        if neighbors[neighbor] in blocks:
-            blocks[neighbors[neighbor]].update()
+        if neighbors[neighbor] in blocks: # !
+            blocks[neighbors[neighbor]].update() # !
 
 def is_occupied(pos):
     pos = inttup(pos)
-    if not pygame.Rect(VEC(pos)*BLOCK_SIZE, (BLOCK_SIZE, BLOCK_SIZE)).colliderect(pygame.Rect(player.pos, player.size)):
+    if not pygame.Rect(VEC(pos) * BLOCK_SIZE, (BLOCK_SIZE, BLOCK_SIZE)).colliderect(pygame.Rect(player.pos, player.size)): # !
         if pos in blocks:
-            return not "replaceable" in blocks[pos].data
+            return not "replaceable" in blocks[pos].data # !
         else:
             return False
     return True
@@ -84,8 +84,8 @@ def is_supported(pos, data, neighbors, c=False):
         supports = data["support"]
         for support in supports:
             if inttup(support.split(" ")) != inttup(VEC(c)-VEC(pos)):
-                if neighbors[support] in blocks:
-                    if blocks[neighbors[support]].name not in supports[support]:
+                if neighbors[support] in blocks: # !
+                    if blocks[neighbors[support]].name not in supports[support]: # !
                         return False
                 else:
                     return False
