+++ src/constants.py
@@ -11,6 +11,7 @@ WIDTH, HEIGHT = 1200, 600
 SCR_DIM = (WIDTH, HEIGHT)
 GRAVITY = 0.5
 SLIDE = 0.3
+SPACING = 24
 TERMINAL_VEL = 24
 BLOCK_SIZE = 64
 CHUNK_SIZE = 8
